requirements.txt
streamlit
langchain_huggingface
langchain_text_splitters
langchain
langchain_core
langchain_community
python-dotenv
langchain[groq]
ipykernel 
bs4
arxiv
pymupdf
chromadb
sentence-transformers
faiss-cpu
pandas
langgraph
pypdf